requirements.txt
pandas==0.23.4
exotoolbox==0.1
astroquery==0.3.9
objgraph==3.4.0
tqdm==4.26.0
ccdproc==1.3.0.post1
matplotlib==2.2.3
numpy==1.15.1
pymultinest==2.7
psutil==5.4.7
astropy==3.0.4
jdcal==1.4
seaborn==0.9.0
photutils==0.6
scipy==1.1.0
george==0.3.1
batman_package==2.4.6
PyQt5==5.11.3
batman==0.7.0
python_dateutil==2.7.5
pyephem==3.7.6.0
mysql_connector_repackaged==0.3.1
